python/requirements.txt
pyfakefs==5.9.1
jpype1==1.6.0
packaging==24.1
